requirements.txt
streamlit
sentence-transformers
pinecone
python-dotenv
numpy